rename lf1 to lf, in preparation for ABC
