+++ floodmodeller_api/logs/lf1_params.py
@@ -14,7 +14,7 @@ If you have any query about this program or this License, please contact us at s
 address: Jacobs UK Limited, Flood Modeller, Cottons Centre, Cottons Lane, London, SE1 2QG, United Kingdom.
 """
 
-from .lf1_helpers import (
+from .lf_helpers import (
     DateTime,
     Time,
     TimeDeltaHMS,
